pyproject.toml
[tool.poetry]
name = "mailing_service"
version = "0.1.0"
description = ""
authors = ["kill2002@mail.ru"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
django-redis = "^5.3.0"
django-crontab = "^0.7.1"
python-crontab = "^3.0.0"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
